+++ Assets/Scripts/Constraints/VLineConstraint.cs
@@ -1,66 +1,65 @@
 ﻿using UnityEngine;
 
 
-public class VLineConstraint :Constraint 
+public class VLineConstraint :Constraint
 {
 
-	private BlockSparseMatrix.MatrixBlock m_MatrixBlockJ;
-	private BlockSparseMatrix.MatrixBlock m_MatrixBlockJDot;
-	private readonly Particle m_Particle;
-	private float m_X;
+    private BlockSparseMatrix.MatrixBlock m_MatrixBlockJ;
+    private BlockSparseMatrix.MatrixBlock m_MatrixBlockJDot;
+    private readonly Particle m_Particle;
+    private double m_X;
 
-	public VLineConstraint(Particle a_Particle, ParticleSystem a_System)
-	{
-		m_Particle = a_Particle;
-		m_X = a_Particle.Position.x;
-		// We can probably de-duplicate this code:
-		int j = a_System.GetParticleIndex(a_Particle) * a_System.GetParticleDimension();
-		int i = a_System.AddConstraint(this);
+    public VLineConstraint(Particle a_Particle, ParticleSystem a_System)
+    {
+        m_Particle = a_Particle;
+        m_X = a_Particle.Position.x;
+        // We can probably de-duplicate this code:
+        int j = a_System.GetParticleIndex(a_Particle) * a_System.GetParticleDimension();
+        int i = a_System.AddConstraint(this);
 
-		Debug.Log("Creating VLine constraint with index " + i);
-		int iLength = GetConstraintDimension();
-		int jLength = a_System.GetParticleDimension();
-		m_MatrixBlockJ = a_System.MatrixJ.CreateMatrixBlock(i, j, iLength, jLength);
-		m_MatrixBlockJDot = a_System.MatrixJDot.CreateMatrixBlock(i, j, iLength, jLength);
-	}
+        Debug.Log("Creating VLine constraint with index " + i);
+        int iLength = GetConstraintDimension();
+        int jLength = a_System.GetParticleDimension();
+        m_MatrixBlockJ = a_System.MatrixJ.CreateMatrixBlock(i, j, iLength, jLength);
+        m_MatrixBlockJDot = a_System.MatrixJDot.CreateMatrixBlock(i, j, iLength, jLength);
+    }
 
-	public void UpdateJacobians(ParticleSystem a_ParticleSystem)
-	{
-		
-		m_MatrixBlockJ.data[0] = 1;
-		m_MatrixBlockJ.data[1] = 0;
-		//Jdot = 0
-		m_MatrixBlockJDot.data[0] = 0;
-		m_MatrixBlockJDot.data[1] = 0; 
-	}
+    public void UpdateJacobians(ParticleSystem a_ParticleSystem)
+    {
+        m_MatrixBlockJ.data[0] = 1;
+        m_MatrixBlockJ.data[1] = 0;
+        //Jdot = 0
+        m_MatrixBlockJDot.data[0] = 0;
+        m_MatrixBlockJDot.data[1] = 0; 
+    }
 
-	public float[] GetValue(ParticleSystem a_ParticleSystem)
-	{
-		float[] v = new float[GetConstraintDimension ()];
-		v [0] = m_Particle.Position.x - m_X;
-		return v;
-	}
+    public double[] GetValue(ParticleSystem a_ParticleSystem)
+    {
+        double[] v = new double[GetConstraintDimension()];
+        v[0] = m_Particle.Position.x - m_X;
+        return v;
+    }
 
-	public float[] GetDerivativeValue(ParticleSystem a_ParticleSystem)
-	{
-		float[] v = new float[GetConstraintDimension ()];
-		v [0] = m_Particle.Velocity.x;
-		return v;
-	}
+    public double[] GetDerivativeValue(ParticleSystem a_ParticleSystem)
+    {
+        double[] v = new double[GetConstraintDimension()];
+        v[0] = m_Particle.Velocity.x;
+        return v;
+    }
 
-	public int GetConstraintDimension()
-	{
-		return 1;
-	}
+    public int GetConstraintDimension()
+    {
+        return 1;
+    }
 
-	public void Draw()
-	{
-		GL.Begin(GL.LINES);
-		GL.Color(new Color(1f, 0.5f, 0.5f));
-		GL.Vertex(new Vector2(m_X,-100));
-		GL.Vertex(new Vector2(m_X,100));
-		GL.End();
-	}
+    public void Draw()
+    {
+        GL.Begin(GL.LINES);
+        GL.Color(new Color(1f, 0.5f, 0.5f));
+        GL.Vertex(new Vector2((float)m_X, -100));
+        GL.Vertex(new Vector2((float)m_X, 100));
+        GL.End();
+    }
 }
